requirements.txt
# Web scraping
aiohttp>=3.8.0
beautifulsoup4>=4.12.0

# NLP
spacy>=3.6.0
vaderSentiment>=3.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Kafka
confluent-kafka>=2.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.0
websockets>=12.0

# Orchestration
prefect>=2.14.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0

# dbt (optional)
dbt-core>=1.7.0
dbt-postgres>=1.7.0

# Monitoring
prometheus-client>=0.19.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
